Feat: home onboarding add animation
