+++ src/components/Home/Onboarding/Icons.tsx
@@ -1,0 +1,90 @@
+const closeButton = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+  >
+    <path
+      d="M13.2142 12L19.3666 4.66641C19.4697 4.54453 19.383 4.35938 19.2236 4.35938L17.3533 4.35938C17.2432 4.35938 17.1377 4.40859 17.065 4.49297L11.9908 10.5422L6.91659 4.49297C6.84628 4.40859 6.74081 4.35938 6.62831 4.35938L4.758 4.35938C4.59863 4.35938 4.51191 4.54453 4.61503 4.66641L10.7674 12L4.61503 19.3336C4.59193 19.3608 4.57711 19.394 4.57233 19.4293C4.56755 19.4647 4.57301 19.5006 4.58806 19.533C4.60312 19.5653 4.62713 19.5926 4.65725 19.6117C4.68737 19.6308 4.72234 19.6408 4.758 19.6406L6.62831 19.6406C6.73847 19.6406 6.84394 19.5914 6.91659 19.507L11.9908 13.4578L17.065 19.507C17.1353 19.5914 17.2408 19.6406 17.3533 19.6406L19.2236 19.6406C19.383 19.6406 19.4697 19.4555 19.3666 19.3336L13.2142 12Z"
+      fill="#23272F"
+    />
+  </svg>
+);
+const fireIcon = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+  >
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M12.3945 2.55279C12.2662 2.29624 12.034 2.10713 11.7568 2.03351C11.4795 1.95988 11.184 2.00885 10.9454 2.16795C10.5995 2.39858 10.3314 2.72608 10.1229 3.04791C9.90855 3.37854 9.71986 3.76148 9.553 4.16366C9.21939 4.96773 8.93911 5.93195 8.71375 6.89778C8.42752 8.12448 8.21568 9.41687 8.10004 10.4776C7.61585 10.1512 7.33491 9.78527 7.15481 9.41104C6.82729 8.73046 6.75736 7.8772 6.75736 6.75739C6.75736 6.35292 6.51372 5.98829 6.14004 5.83351C5.76637 5.67872 5.33625 5.76428 5.05025 6.05028C3.68361 7.41692 3 9.21013 3 11C3 12.7899 3.68361 14.5831 5.05025 15.9498C7.78392 18.6834 12.2161 18.6834 14.9497 15.9498C16.3164 14.5831 17 12.7899 17 11C17 9.21013 16.3164 7.41692 14.9497 6.05028C14.3584 5.45889 13.9696 5.06453 13.6021 4.5828C13.239 4.10688 12.8781 3.51991 12.3945 2.55279ZM12.1213 15.1213C10.9497 16.2929 9.05025 16.2929 7.87868 15.1213C7.29289 14.5355 7 13.7678 7 13C7 13 7.87868 13.5 9.50005 13.5C9.50005 12.5 10 9.5 10.75 9C11.25 10 11.5355 10.2929 12.1213 10.8787C12.7071 11.4645 13 12.2322 13 13C13 13.7678 12.7071 14.5355 12.1213 15.1213Z"
+      fill="#FF85B1"
+    />
+  </svg>
+);
+const snowIcon = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+  >
+    <path
+      d="M10.0012 20C9.83549 20 9.67651 19.9342 9.5593 19.8169C9.44209 19.6997 9.37625 19.5408 9.37625 19.375V17.7587L8.56875 18.5675C8.45139 18.6847 8.29228 18.7505 8.12643 18.7503C7.96058 18.7502 7.80156 18.6842 7.68437 18.5669C7.56718 18.4495 7.50141 18.2904 7.50152 18.1246C7.50164 17.9587 7.56764 17.7997 7.685 17.6825L9.37625 15.9913V14.6337L8.56875 15.4425C8.51064 15.5005 8.44167 15.5465 8.36577 15.5779C8.28988 15.6093 8.20855 15.6254 8.12643 15.6253C7.96058 15.6252 7.80156 15.5592 7.68437 15.4419C7.56718 15.3245 7.50141 15.1654 7.50152 14.9996C7.50164 14.8337 7.56764 14.6747 7.685 14.5575L9.37625 12.8663V11.0825L7.83125 11.9738L7.2125 14.2863C7.19116 14.3655 7.15441 14.4398 7.10435 14.5049C7.0543 14.57 6.99191 14.6246 6.92076 14.6656C6.8496 14.7065 6.77108 14.7331 6.68966 14.7437C6.60824 14.7544 6.52553 14.7488 6.44625 14.7275C6.36696 14.7062 6.29265 14.6694 6.22757 14.6194C6.16249 14.5693 6.1079 14.5069 6.06693 14.4358C6.02595 14.3646 5.99939 14.2861 5.98877 14.2047C5.97814 14.1232 5.98366 14.0405 6.005 13.9612L6.30125 12.8587L5.12625 13.5363L4.50625 15.8488C4.46315 16.0089 4.35821 16.1453 4.21451 16.2281C4.07081 16.3108 3.90012 16.3331 3.74 16.29C3.57987 16.2469 3.44343 16.142 3.36068 15.9983C3.27793 15.8546 3.25565 15.6839 3.29875 15.5237L3.595 14.4212L2.195 15.2287C2.05152 15.3103 1.88161 15.3318 1.72233 15.2886C1.56305 15.2453 1.42735 15.1409 1.34483 14.9979C1.2623 14.855 1.23966 14.6852 1.28185 14.5257C1.32403 14.3661 1.42762 14.2297 1.57 14.1463L2.97 13.3387L1.865 13.0425C1.78571 13.0212 1.71141 12.9844 1.64632 12.9344C1.58124 12.8843 1.52665 12.8219 1.48568 12.7508C1.4447 12.6796 1.41814 12.6011 1.40752 12.5197C1.39689 12.4382 1.40241 12.3555 1.42375 12.2762C1.44509 12.197 1.48183 12.1227 1.53189 12.0576C1.58195 11.9925 1.64433 11.9379 1.71549 11.8969C1.78664 11.856 1.86517 11.8294 1.94658 11.8188C2.028 11.8081 2.11071 11.8137 2.19 11.835L4.5 12.4538L5.675 11.7762L4.5725 11.48C4.492 11.4599 4.41631 11.4241 4.34983 11.3744C4.28334 11.3248 4.2274 11.2625 4.18527 11.191C4.14313 11.1195 4.11564 11.0404 4.1044 10.9582C4.09317 10.876 4.0984 10.7924 4.11981 10.7123C4.14121 10.6321 4.17836 10.557 4.22908 10.4914C4.2798 10.4258 4.34309 10.3709 4.41524 10.3299C4.48739 10.289 4.56698 10.2628 4.64934 10.253C4.73171 10.2431 4.81522 10.2498 4.895 10.2725L7.2075 10.8913L8.75125 10L7.20625 9.10875L4.895 9.7275C4.81522 9.75024 4.73171 9.75688 4.64934 9.74702C4.56698 9.73715 4.48739 9.71099 4.41524 9.67006C4.34309 9.62913 4.2798 9.57424 4.22908 9.50859C4.17836 9.44295 4.14121 9.36787 4.11981 9.28772C4.0984 9.20757 4.09317 9.12397 4.1044 9.04178C4.11564 8.95958 4.14313 8.88045 4.18527 8.80899C4.2274 8.73754 4.28334 8.67518 4.34983 8.62556C4.41631 8.57595 4.492 8.54006 4.5725 8.52L5.67625 8.22375L4.50125 7.54625L2.18875 8.165C2.02966 8.20465 1.86137 8.1802 1.72015 8.09693C1.57892 8.01365 1.47607 7.87822 1.43377 7.71982C1.39146 7.56143 1.41309 7.39275 1.49399 7.25015C1.57489 7.10755 1.70858 7.00245 1.86625 6.9575L2.97 6.66125L1.57 5.85375C1.49834 5.81301 1.43546 5.7585 1.38495 5.69336C1.33445 5.62822 1.29734 5.55373 1.27574 5.47418C1.25414 5.39464 1.2485 5.31161 1.25912 5.22987C1.26974 5.14813 1.29643 5.0693 1.33764 4.99792C1.37886 4.92654 1.43378 4.86401 1.49926 4.81395C1.56474 4.76388 1.63947 4.72726 1.71916 4.70619C1.79884 4.68512 1.88191 4.68002 1.96357 4.69119C2.04524 4.70235 2.12389 4.72956 2.195 4.77125L3.595 5.57875L3.29875 4.475C3.2591 4.31592 3.28355 4.14763 3.36682 4.0064C3.45009 3.86517 3.58553 3.76232 3.74392 3.72002C3.90232 3.67771 4.071 3.69934 4.2136 3.78024C4.3562 3.86114 4.4613 3.99483 4.50625 4.1525L5.125 6.46375L6.3 7.14125L6.005 6.0375C5.96535 5.87842 5.9898 5.71013 6.07307 5.5689C6.15634 5.42767 6.29178 5.32482 6.45017 5.28252C6.60857 5.24021 6.77725 5.26184 6.91985 5.34274C7.06244 5.42364 7.16755 5.55733 7.2125 5.715L7.83125 8.02625L9.37625 8.9175V7.13375L7.685 5.4425C7.56764 5.32531 7.50164 5.16629 7.50152 5.00044C7.50141 4.83459 7.56718 4.67548 7.68437 4.55813C7.80156 4.44077 7.96058 4.37477 8.12643 4.37465C8.29228 4.37454 8.45139 4.44031 8.56875 4.5575L9.37625 5.36625V4.00875L7.685 2.3175C7.62689 2.25947 7.58077 2.19057 7.54929 2.11472C7.51781 2.03887 7.50158 1.95756 7.50152 1.87544C7.50147 1.79332 7.51758 1.71199 7.54896 1.6361C7.58033 1.56021 7.62634 1.49123 7.68437 1.43313C7.7424 1.37502 7.8113 1.3289 7.88715 1.29742C7.963 1.26594 8.04431 1.24971 8.12643 1.24965C8.20855 1.24959 8.28988 1.26571 8.36577 1.29709C8.44167 1.32846 8.51064 1.37447 8.56875 1.4325L9.37625 2.24125V0.625C9.37625 0.45924 9.44209 0.300269 9.5593 0.183058C9.67651 0.065848 9.83549 0 10.0012 0C10.167 0 10.326 0.065848 10.4432 0.183058C10.5604 0.300269 10.6262 0.45924 10.6262 0.625V2.24125L11.435 1.4325C11.5524 1.31531 11.7115 1.24954 11.8773 1.24965C12.0432 1.24977 12.2022 1.31577 12.3194 1.43313C12.4366 1.55048 12.5023 1.70959 12.5022 1.87544C12.5021 2.04129 12.4361 2.20031 12.3187 2.3175L10.6262 4.00875V5.36625L11.435 4.5575C11.5524 4.44031 11.7115 4.37454 11.8773 4.37465C12.0432 4.37477 12.2022 4.44077 12.3194 4.55813C12.4366 4.67548 12.5023 4.83459 12.5022 5.00044C12.5021 5.16629 12.4361 5.32531 12.3187 5.4425L10.6262 7.13375V8.9175L12.1712 8.02625L12.79 5.71375C12.8113 5.63446 12.8481 5.56016 12.8981 5.49507C12.9482 5.42999 13.0106 5.3754 13.0817 5.33443C13.1529 5.29346 13.2314 5.2669 13.3128 5.25627C13.3942 5.24565 13.477 5.25116 13.5562 5.2725C13.6355 5.29384 13.7098 5.33059 13.7749 5.38064C13.84 5.4307 13.8946 5.49309 13.9356 5.56424C13.9765 5.63539 14.0031 5.71392 14.0137 5.79534C14.0244 5.87675 14.0188 5.95946 13.9975 6.03875L13.7025 7.14125L14.8775 6.46375L15.4962 4.15125C15.5393 3.99113 15.6443 3.85468 15.788 3.77193C15.9317 3.68918 16.1024 3.6669 16.2625 3.71C16.4226 3.7531 16.5591 3.85804 16.6418 4.00174C16.7246 4.14544 16.7468 4.31613 16.7037 4.47625L16.4087 5.57875L17.8087 4.77125C17.9522 4.68968 18.1221 4.66816 18.2814 4.71141C18.4407 4.75465 18.5764 4.85914 18.6589 5.00207C18.7414 5.145 18.7641 5.31476 18.7219 5.47432C18.6797 5.63388 18.5761 5.77028 18.4337 5.85375L17.0337 6.66125L18.1375 6.9575C18.2976 7.0006 18.4341 7.10554 18.5168 7.24924C18.5996 7.39294 18.6218 7.56363 18.5787 7.72375C18.5356 7.88387 18.4307 8.02032 18.287 8.10307C18.1433 8.18582 17.9726 8.2081 17.8125 8.165L15.5025 7.54625L14.3275 8.22375L15.4312 8.52C15.5914 8.5631 15.7278 8.66804 15.8106 8.81174C15.8933 8.95544 15.9156 9.12613 15.8725 9.28625C15.8294 9.44637 15.7245 9.58282 15.5808 9.66557C15.4371 9.74832 15.2664 9.7706 15.1062 9.7275L12.7962 9.10875L11.2512 10L12.7962 10.8913L15.1075 10.2725C15.2676 10.2296 15.4382 10.252 15.5818 10.3349C15.7254 10.4177 15.8302 10.5542 15.8731 10.7144C15.9161 10.8745 15.8936 11.0451 15.8107 11.1887C15.7279 11.3323 15.5914 11.4371 15.4312 11.48L14.3275 11.7762L15.5025 12.4538L17.8137 11.835C17.9739 11.7921 18.1445 11.8145 18.2881 11.8974C18.4317 11.9802 18.5364 12.1167 18.5794 12.2769C18.6223 12.437 18.5999 12.6076 18.517 12.7512C18.4341 12.8948 18.2976 12.9996 18.1375 13.0425L17.0337 13.3387L18.4337 14.1463C18.5761 14.2297 18.6797 14.3661 18.7219 14.5257C18.7641 14.6852 18.7414 14.855 18.6589 14.9979C18.5764 15.1409 18.4407 15.2453 18.2814 15.2886C18.1221 15.3318 17.9522 15.3103 17.8087 15.2287L16.4087 14.4212L16.7037 15.525C16.7265 15.6048 16.7331 15.6883 16.7233 15.7707C16.7134 15.853 16.6872 15.9326 16.6463 16.0048C16.6054 16.0769 16.5505 16.1402 16.4848 16.1909C16.4192 16.2416 16.3441 16.2788 16.264 16.3002C16.1838 16.3216 16.1002 16.3268 16.018 16.3156C15.9358 16.3044 15.8567 16.2769 15.7852 16.2347C15.7138 16.1926 15.6514 16.1367 15.6018 16.0702C15.5522 16.0037 15.5163 15.928 15.4962 15.8475L14.8775 13.5363L13.7025 12.8587L13.9975 13.9625C14.0371 14.1216 14.0127 14.2899 13.9294 14.4311C13.8461 14.5723 13.7107 14.6752 13.5523 14.7175C13.3939 14.7598 13.2252 14.7382 13.0826 14.6573C12.94 14.5764 12.8349 14.4427 12.79 14.285L12.1712 11.975L10.6262 11.0825V12.8663L12.3187 14.5575C12.4361 14.6747 12.5021 14.8337 12.5022 14.9996C12.5023 15.1654 12.4366 15.3245 12.3194 15.4419C12.2022 15.5592 12.0432 15.6252 11.8773 15.6253C11.7115 15.6255 11.5524 15.5597 11.435 15.4425L10.6262 14.6337V15.9913L12.3187 17.6825C12.4361 17.7997 12.5021 17.9587 12.5022 18.1246C12.5023 18.2904 12.4366 18.4495 12.3194 18.5669C12.2022 18.6842 12.0432 18.7502 11.8773 18.7503C11.7115 18.7505 11.5524 18.6847 11.435 18.5675L10.6262 17.7587V19.375C10.6262 19.5408 10.5604 19.6997 10.4432 19.8169C10.326 19.9342 10.167 20 10.0012 20Z"
+      fill="#979C9E"
+    />
+  </svg>
+);
+const rank1 = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="30"
+    viewBox="0 0 24 30"
+    fill="none"
+  >
+    <path
+      d="M-3.10839e-10 3C-1.39167e-10 1.34315 1.34315 1.39167e-10 3 3.10839e-10L21 2.17587e-09C22.6569 2.34754e-09 24 1.34315 24 3L24 22L12 30L-2.27949e-09 22L-3.10839e-10 3Z"
+      fill="#E23774"
+    />
+    <path
+      d="M20.0707 8.08789L20.0707 18.9359L18.8347 18.9359L18.8347 8.08789L20.0707 8.08789ZM14.6347 13.1039C13.0147 13.1039 11.8387 12.1679 11.8387 10.8239C11.8387 9.49189 13.0147 8.55589 14.6347 8.55589C16.2547 8.55589 17.4307 9.49189 17.4307 10.8239C17.4307 12.1679 16.2547 13.1039 14.6347 13.1039ZM14.6347 9.57589C13.6747 9.57589 13.0267 10.0559 13.0267 10.8239C13.0267 11.5919 13.6747 12.0839 14.6347 12.0839C15.5947 12.0839 16.2427 11.5919 16.2427 10.8239C16.2427 10.0559 15.5947 9.57589 14.6347 9.57589ZM18.2107 14.4239C17.2507 14.6159 16.2667 14.7119 15.3067 14.7959L15.3067 18.5999L14.0707 18.5999L14.0707 14.8679C13.0507 14.9039 12.1147 14.9159 11.2507 14.9159L11.1067 13.8959C13.0507 13.8959 15.7987 13.8599 18.1267 13.5119L18.2107 14.4239Z"
+      fill="white"
+    />
+    <path
+      d="M5.36225 12.0479L4.63025 11.1599L7.03025 9.10791L8.11025 9.10791L8.11025 16.8479L9.77825 16.8479L9.77825 17.9999L4.94225 17.9999L4.94225 16.8479L6.79025 16.8479L6.79025 11.4359C6.79025 11.2439 6.79025 11.0399 6.80225 10.8359L5.36225 12.0479Z"
+      fill="white"
+    />
+  </svg>
+);
+const rank2 = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="30"
+    viewBox="0 0 24 30"
+    fill="none"
+  >
+    <path
+      d="M-3.10839e-10 3C-1.39167e-10 1.34315 1.34315 1.39167e-10 3 3.10839e-10L21 2.17587e-09C22.6569 2.34754e-09 24 1.34315 24 3L24 22L12 30L-2.27949e-09 22L-3.10839e-10 3Z"
+      fill="#2388FF"
+    />
+    <path
+      d="M20.0695 8.08789L20.0695 18.9359L18.8335 18.9359L18.8335 8.08789L20.0695 8.08789ZM14.6335 13.1039C13.0135 13.1039 11.8375 12.1679 11.8375 10.8239C11.8375 9.49189 13.0135 8.55589 14.6335 8.55589C16.2535 8.55589 17.4295 9.49189 17.4295 10.8239C17.4295 12.1679 16.2535 13.1039 14.6335 13.1039ZM14.6335 9.57589C13.6735 9.57589 13.0255 10.0559 13.0255 10.8239C13.0255 11.5919 13.6735 12.0839 14.6335 12.0839C15.5935 12.0839 16.2415 11.5919 16.2415 10.8239C16.2415 10.0559 15.5935 9.57589 14.6335 9.57589ZM18.2095 14.4239C17.2495 14.6159 16.2655 14.7119 15.3055 14.7959L15.3055 18.5999L14.0695 18.5999L14.0695 14.8679C13.0495 14.9039 12.1135 14.9159 11.2495 14.9159L11.1055 13.8959C13.0495 13.8959 15.7975 13.8599 18.1255 13.5119L18.2095 14.4239Z"
+      fill="white"
+    />
+    <path
+      d="M10.0293 18.0001L4.0293 18.0001L4.0293 17.2321L7.0173 14.2321C7.4253 13.8241 7.8093 13.4161 8.1213 12.9241C8.7213 11.8441 8.7933 10.1641 7.0893 10.1641C6.1413 10.1641 5.4333 10.8601 5.5053 11.8081L4.1853 11.5681C4.4133 9.98413 5.5173 9.03613 7.1613 9.03613C8.6493 9.03613 9.8733 9.92413 9.8733 11.5321C9.8733 12.8761 8.9493 13.8241 8.0613 14.7121L5.8053 16.9921C6.1773 16.8961 6.5733 16.8241 6.9573 16.8241L10.0293 16.8241L10.0293 18.0001Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export { closeButton, fireIcon, rank1, rank2, snowIcon };
